requirements.txt
apkverify
pyaxmlparser
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
-e git+https://github.com/Unbrick/jodel_api.git@be12ddcabd59b9fcd9ad08c1ff703a05b1d08b00#egg=jodel_api
MarkupSafe==1.1.0
mock==2.0.0
pbr==5.1.1
protobuf==3.6.1
pycodestyle==2.4.0
python-magic-bin==0.4.14
r2pipe==1.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
varint==1.0.2
Werkzeug==0.14.1
